+++ posthog-core/test/posthog.featureflags.v3.spec.ts
@@ -37,7 +37,7 @@ describe('PostHog Feature Flags v3', () => {
   beforeEach(() => {
     ;[posthog, mocks] = createTestClient('TEST_API_KEY', { flushAt: 1 }, (_mocks) => {
       _mocks.fetch.mockImplementation((url) => {
-        if (url.includes('/decide/?v=4')) {
+        if (url.includes('/flags/?v=2')) {
           return Promise.resolve({
             status: 200,
             text: () => Promise.resolve('ok'),
@@ -135,7 +135,7 @@ describe('PostHog Feature Flags v3', () => {
         beforeEach(() => {
           ;[posthog, mocks] = createTestClient('TEST_API_KEY', { flushAt: 1 }, (_mocks) => {
             _mocks.fetch.mockImplementation((url) => {
-              if (url.includes('/decide/')) {
+              if (url.includes('/flags/')) {
                 return Promise.resolve({
                   status: 400,
                   text: () => Promise.resolve('ok'),
@@ -154,7 +154,7 @@ describe('PostHog Feature Flags v3', () => {
         })
 
         it('should return undefined', async () => {
-          expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/decide/?v=4', {
+          expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/flags/?v=2', {
             body: JSON.stringify({
               token: 'TEST_API_KEY',
               distinct_id: posthog.getDistinctId(),
@@ -189,7 +189,7 @@ describe('PostHog Feature Flags v3', () => {
           ;[posthog, mocks] = createTestClient('TEST_API_KEY', { flushAt: 1 }, (_mocks) => {
             _mocks.fetch
               .mockImplementationOnce((url) => {
-                if (url.includes('/decide/?v=4')) {
+                if (url.includes('/flags/?v=2')) {
                   return Promise.resolve({
                     status: 200,
                     text: () => Promise.resolve('ok'),
@@ -202,7 +202,7 @@ describe('PostHog Feature Flags v3', () => {
                 return errorAPIResponse
               })
               .mockImplementationOnce((url) => {
-                if (url.includes('/decide/?v=4')) {
+                if (url.includes('/flags/?v=2')) {
                   return Promise.resolve({
                     status: 200,
                     text: () => Promise.resolve('ok'),
@@ -225,7 +225,7 @@ describe('PostHog Feature Flags v3', () => {
         })
 
         it('should return combined results', async () => {
-          expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/decide/?v=4', {
+          expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/flags/?v=2', {
             body: JSON.stringify({
               token: 'TEST_API_KEY',
               distinct_id: posthog.getDistinctId(),
@@ -252,7 +252,7 @@ describe('PostHog Feature Flags v3', () => {
           // now second call to feature flags
           await posthog.reloadFeatureFlagsAsync(false)
 
-          expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/decide/?v=4', {
+          expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/flags/?v=2', {
             body: JSON.stringify({
               token: 'TEST_API_KEY',
               distinct_id: posthog.getDistinctId(),
@@ -294,7 +294,7 @@ describe('PostHog Feature Flags v3', () => {
           ;[posthog, mocks] = createTestClient('TEST_API_KEY', { flushAt: 1 }, (_mocks) => {
             _mocks.fetch
               .mockImplementationOnce((url) => {
-                if (url.includes('/decide/?v=4')) {
+                if (url.includes('/flags/?v=2')) {
                   return Promise.resolve({
                     status: 200,
                     text: () => Promise.resolve('ok'),
@@ -307,7 +307,7 @@ describe('PostHog Feature Flags v3', () => {
                 return errorAPIResponse
               })
               .mockImplementationOnce((url) => {
-                if (url.includes('/decide/?v=4')) {
+                if (url.includes('/flags/?v=2')) {
                   return Promise.resolve({
                     status: 200,
                     text: () => Promise.resolve('ok'),
@@ -330,7 +330,7 @@ describe('PostHog Feature Flags v3', () => {
         })
 
         it('should return only latest results', async () => {
-          expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/decide/?v=4', {
+          expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/flags/?v=2', {
             body: JSON.stringify({
               token: 'TEST_API_KEY',
               distinct_id: posthog.getDistinctId(),
@@ -357,7 +357,7 @@ describe('PostHog Feature Flags v3', () => {
           // now second call to feature flags
           await posthog.reloadFeatureFlagsAsync(false)
 
-          expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/decide/?v=4', {
+          expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/flags/?v=2', {
             body: JSON.stringify({
               token: 'TEST_API_KEY',
               distinct_id: posthog.getDistinctId(),
@@ -553,7 +553,7 @@ describe('PostHog Feature Flags v3', () => {
       beforeEach(() => {
         ;[posthog, mocks] = createTestClient('TEST_API_KEY', { flushAt: 1 }, (_mocks) => {
           _mocks.fetch.mockImplementation((url) => {
-            if (url.includes('/decide/')) {
+            if (url.includes('/flags/')) {
               return Promise.resolve({
                 status: 200,
                 text: () => Promise.resolve('ok'),
@@ -574,7 +574,7 @@ describe('PostHog Feature Flags v3', () => {
 
       it('should unset all flags when feature_flags is quota limited', async () => {
         // First verify the fetch was called correctly
-        expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/decide/?v=4', {
+        expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/flags/?v=2', {
           body: JSON.stringify({
             token: 'TEST_API_KEY',
             distinct_id: posthog.getDistinctId(),
@@ -640,7 +640,7 @@ describe('PostHog Feature Flags v3', () => {
         },
         (_mocks) => {
           _mocks.fetch.mockImplementation((url) => {
-            if (url.includes('/decide/')) {
+            if (url.includes('/flags/')) {
               return Promise.reject(new Error('Not responding to emulate use of bootstrapped values'))
             }
 
@@ -748,7 +748,7 @@ describe('PostHog Feature Flags v3', () => {
           },
           (_mocks) => {
             _mocks.fetch.mockImplementation((url) => {
-              if (url.includes('/decide/')) {
+              if (url.includes('/flags/')) {
                 return Promise.resolve({
                   status: 200,
                   text: () => Promise.resolve('ok'),
@@ -776,7 +776,7 @@ describe('PostHog Feature Flags v3', () => {
       })
 
       it('should load new feature flags', async () => {
-        expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/decide/?v=4', {
+        expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/flags/?v=2', {
           body: JSON.stringify({
             token: 'TEST_API_KEY',
             distinct_id: posthog.getDistinctId(),
@@ -802,7 +802,7 @@ describe('PostHog Feature Flags v3', () => {
       })
 
       it('should load new feature flag payloads', async () => {
-        expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/decide/?v=4', {
+        expect(mocks.fetch).toHaveBeenCalledWith('https://us.i.posthog.com/flags/?v=2', {
           body: JSON.stringify({
             token: 'TEST_API_KEY',
             distinct_id: posthog.getDistinctId(),
@@ -870,7 +870,7 @@ describe('PostHog Feature Flags v3', () => {
         },
         (_mocks) => {
           _mocks.fetch.mockImplementation((url) => {
-            if (url.includes('/decide/')) {
+            if (url.includes('/flags/')) {
               return Promise.resolve({
                 status: 200,
                 text: () => Promise.resolve('ok'),